move peek_definition to extensions.lua
